Use LXQtCompilerSettings
It will be increasingly harder to use them. So, the sooner the better.

clazy was used to help in the process of porting from automatic conversions
from 8-bit strings (char *) to unicode QStrings. That was the main barrier.
There's room for improvement for sure.

The foreach Qt commit will be handled in another commit.