play sound
The deployed version of organizer 0.6.0 is not playing the audio file embedded in the resource sounds.qrc.

Changed audio source so that it links directly to a local freedesktop sound file.

[no sound on linux](https://forum.qt.io/topic/54162/solved-qapplication-beep-no-sound-on-linux/8)

[fix]: play audio (now from local file)

[minor:] further string changes using macros

[QStringLiteral vs QLatin1String](https://forum.qt.io/topic/78540/qstringliteral-vs-qlatin1string)

[QtDocs: More efficient string contruction say:](https://doc.qt.io/qt-5/qstring.html#more-efficient-string-construction)

Many strings are known at compile time. But the trivial constructor QString("Hello"), will copy the contents of the string, treating the contents as Latin-1. To avoid this one can use the QStringLiteral macro to directly create the required data at compile time. Constructing a QString out of the literal does then not cause any overhead at runtime. A slightly less efficient way is to use QLatin1String.

Assuming QStringLiteral should be used if a QString needs to be created.

[minor:] changing calendar font size with keys f(up) d(down) and r(reset) to partially address issue

[How can I prevent the font from scaling up when the DPI is increased?](https://wiki.qt.io/Technical_FAQ#How_can_I_add_SSL_support_to_my_Qt_application.3F)

On branch master
Changes to be committed:
	renamed:    screenshots/v005-new-appointment.png -> screenshots/v050-new-appointment.png
	renamed:    screenshots/v005-new-contact.png -> screenshots/v050-new-contact.png
	renamed:    screenshots/organizer-v005.png -> screenshots/v050-organizer.png
	renamed:    screenshots/v005-repeat-appointment.png -> screenshots/v050-repeat-appointment.png
	renamed:    screenshots/v005-update-appointment.png -> screenshots/v050-update-appointment.png
	new file:   screenshots/v060-reminder.png
	modified:   src/birthdaylistmodel.cpp
	modified:   src/contactmodel.cpp
	modified:   src/dialogabout.cpp
	modified:   src/dialogappointment.h
	modified:   src/dialogappointmentupdate.cpp
	modified:   src/dialogappointmentupdate.h
	modified:   src/dialogpreferences.h
	modified:   src/mainwindow.cpp
	modified:   src/mainwindow.h